Change type-only imports to type X = import("./x").X
