+++ docs/lib/webglutenfree.d.ts
@@ -1,9 +1,9 @@
 export { BufferBits, BufferUsage, DataType, Primitive } from "./types";
-export { Device, Extension, DeviceCreateOptions, DeviceWithCanvasOptions, DeviceWithContextOptions } from "./device";
+export { Device, Extension, DeviceCreateOptions, DeviceWithCanvasOptions, DeviceWithContextOptions, } from "./device";
 export { TargetClearOptions } from "./target";
-export { Command, CommandOptions, DepthFunc, StencilFunc, StencilOp, BlendFunc, BlendEquation, StencilOrSeparate, BlendOrSeparate, Accessor, Uniform1f, Uniform1fv, Uniform1i, Uniform1iv, Uniform1ui, Uniform1uiv, Uniform2f, Uniform2fv, Uniform2i, Uniform2iv, Uniform2ui, Uniform2uiv, Uniform3f, Uniform3fv, Uniform3i, Uniform3iv, Uniform3ui, Uniform3uiv, Uniform4f, Uniform4fv, Uniform4i, Uniform4iv, Uniform4ui, Uniform4uiv, UniformMatrix2fv, UniformMatrix3fv, UniformMatrix4fv, Uniform, Uniforms, TextureAccessor, Textures } from "./command";
-export { VertexBuffer, VertexBufferType, VertexBufferOptions, VertexBufferStoreOptions } from "./vertex-buffer";
-export { ElementBuffer, ElementBufferType, ElementBufferOptions, ElementBufferStoreOptions } from "./element-buffer";
-export { Attributes, AttributeType, AttributesCreateOptions } from "./attributes";
-export { Texture, TextureFilter, TextureMinFilter, TextureMagFilter, TextureWrap, TextureInternalFormat, TextureFormat, TextureDataType, TextureOptions, TextureStoreOptions } from "./texture";
-export { Framebuffer } from "./framebuffer";
+export { Command, CommandOptions, DepthFunc, StencilFunc, StencilOp, BlendFunc, BlendEquation, StencilOrSeparate, BlendOrSeparate, Accessor, Uniform1f, Uniform1fv, Uniform1i, Uniform1iv, Uniform1ui, Uniform1uiv, Uniform2f, Uniform2fv, Uniform2i, Uniform2iv, Uniform2ui, Uniform2uiv, Uniform3f, Uniform3fv, Uniform3i, Uniform3iv, Uniform3ui, Uniform3uiv, Uniform4f, Uniform4fv, Uniform4i, Uniform4iv, Uniform4ui, Uniform4uiv, UniformMatrix2fv, UniformMatrix3fv, UniformMatrix4fv, Uniform, Uniforms, TextureAccessor, Textures, } from "./command";
+export { VertexBuffer, VertexBufferType, VertexBufferOptions, VertexBufferStoreOptions, } from "./vertex-buffer";
+export { ElementBuffer, ElementBufferType, ElementBufferOptions, ElementBufferStoreOptions, } from "./element-buffer";
+export { Attributes, AttributeType, AttributesCreateOptions, } from "./attributes";
+export { Texture, TextureFilter, TextureMinFilter, TextureMagFilter, TextureWrap, TextureInternalFormat, TextureFormat, TextureDataType, TextureOptions, TextureStoreOptions, } from "./texture";
+export { Framebuffer, TextureColorInternalFormat, TextureDepthInternalFormat, TextureDepthStencilInternalFormat, } from "./framebuffer";
