+++ package.json
@@ -56,7 +56,7 @@
     "rollup-plugin-replace": "^2.0.0",
     "rollup-plugin-uglify": "^3.0.0",
     "tslint": "^5.9.1",
-    "typescript": "^2.8.1",
+    "typescript": "^2.9.0-dev.20180403",
     "uglify-es": "^3.3.8"
   }
 }
