+++ src/target.ts
@@ -1,13 +1,12 @@
 import * as assert from "./util/assert";
 import { BufferBits, Primitive } from "./types";
-import { Device as _Device } from "./device";
-import {
-    Command as _Command,
-    UniformDescriptor as _UniformDescriptor,
-    TextureAccessor,
-} from "./command";
-import { Attributes as _Attributes } from "./attributes";
-import { Framebuffer as _Framebuffer } from "./framebuffer";
+
+export type Device = import ("./device").Device;
+export type Command<P> = import ("./command").Command<P>;
+export type UniformDescriptor<P> = import ("./command").UniformDescriptor<P>;
+export type TextureAccessor<P> = import ("./command").TextureAccessor<P>;
+export type Attributes = import ("./attributes").Attributes;
+export type Framebuffer = import ("./framebuffer").Framebuffer;
 
 export interface TargetClearOptions {
     r?: number;
@@ -25,7 +24,7 @@ export interface TargetClearOptions {
 export class Target {
 
     constructor(
-        private dev: _Device,
+        private dev: Device,
         readonly glDrawBuffers: number[],
         readonly glFramebuffer: WebGLFramebuffer | null,
         readonly width?: number,
@@ -65,7 +64,7 @@ export class Target {
      * Blit source framebuffer onto this render target. Use buffer bits to
      * choose, which buffers to blit.
      */
-    blit(source: _Framebuffer, bits: BufferBits): void {
+    blit(source: Framebuffer, bits: BufferBits): void {
         const {
             dev: { _gl, _stackReadFramebuffer },
             width,
@@ -115,7 +114,7 @@ export class Target {
      * The properties are passed to the command's uniform or texture callbacks,
      * if used.
      */
-    draw<P>(cmd: _Command<P>, attrs: _Attributes, props: P): void {
+    draw<P>(cmd: Command<P>, attrs: Attributes, props: P): void {
         const {
             dev: {
                 _stackVertexArray,
@@ -180,8 +179,8 @@ export class Target {
      * unnecesasry rebinding.
      */
     batch<P>(
-        cmd: _Command<P>,
-        cb: (draw: (attrs: _Attributes, props: P) => void) => void,
+        cmd: Command<P>,
+        cb: (draw: (attrs: Attributes, props: P) => void) => void,
     ): void {
         const {
             dev: {
@@ -214,7 +213,7 @@ export class Target {
 
         let iter = 0;
 
-        cb((attrs: _Attributes, props: P) => {
+        cb((attrs: Attributes, props: P) => {
             // with() ensures the original target is still bound
             this.with(() => {
                 iter++;
@@ -325,7 +324,7 @@ export class Target {
     }
 
     private uniforms<P>(
-        uniformDescrs: _UniformDescriptor<P>[],
+        uniformDescrs: UniformDescriptor<P>[],
         props: P,
         index: number,
     ): void {
